requirements.txt
gradio
requests
PyMuPDF
python-dotenv